feat: delete topics an queues implemented

+++ server/controllers/queue_controller.py
@@ -2,13 +2,15 @@ from fastapi import HTTPException
 from database import insert_queue, find_queue, find_all_queues, update_queue, delete_queue
 from models import QueueModel
 from utils import verify_token
+from state import active_sessions
 
 
 def create_queue(queue: QueueModel, token: str):
     verify_token(token)
+    client = active_sessions[token]
     if find_queue(queue.name):
         raise HTTPException(status_code=400, detail="Queue already exists")
-    insert_queue({"name": queue.name, "messages": []})
+    insert_queue({"name": queue.name, "messages": [], "owner": client})
     print(list(find_all_queues()))
     return {"message": "Queue created"}
 
@@ -39,9 +41,12 @@ def receive_message(queue_name: str, token: str):
 
 def delete_one_queue(queue_name: str, token: str):
     verify_token(token)
+    client = active_sessions[token]
     queue = find_queue(queue_name)
     if not queue:
         raise HTTPException(status_code=404, detail="Queue not found")
-
-    delete_queue(queue_name)
+    if queue["owner"] == client:
+        delete_queue(queue_name)
+    else:
+        return {"message": "You cannot delete this queue"}
     return {"message": "Queue deleted"}
